+++ requirements.txt
@@ -1,0 +1,2 @@
+geopy==1.20.0
+requests==2.22.0
